requirements.txt
opencv-python
python-docx
pywin32
numpy
reportlab
pdfrw
